pyproject.toml
[tool.poetry]
name = "spycatagency"
version = "0.1.0"
description = ""
authors = ["JuliManhupli <juliamangup@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.4"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
django-ninja = "^1.3.0"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
